MCSC-290 Create DC/DB Page Content

+++ app/pages/annualallowance/taxyear/DefinedBenefitAmountPage.scala
@@ -22,7 +22,7 @@ import pages.QuestionPage
 import play.api.libs.json.JsPath
 import play.api.mvc.Call
 
-case class DefinedBenefitAmountPage(period: Period, schemeIndex: SchemeIndex) extends QuestionPage[Int] {
+case class DefinedBenefitAmountPage(period: Period, schemeIndex: SchemeIndex) extends QuestionPage[BigInt] {
 
   override def path: JsPath = JsPath \ "aa" \ "years" \ period.toString \ "schemes" \ schemeIndex.toString \ toString
